requirements.txt
python-telegram-bot==20.3
pymongo==4.5.0
python-dotenv==1.0.0